use module layout
