+++ app/airq/__init__.py
@@ -2,13 +2,13 @@ from airq import config  # Do this first, it initializes everything
 from airq.celery import celery  # This is necesary to start celery
 
 from airq import api
-from airq import commands
+from airq import management
 
 
 app = config.app
 
-# Register commands
-app.cli.command("sync")(commands.sync)
+# Register management commands
+app.cli.command("sync")(management.sync)
 
 # Register routes
 app.route("/", methods=["GET"])(api.healthcheck)
