+++ app/airq/management/__init__.py
@@ -1,0 +1,1 @@
+from airq.management.sync import sync
